+++ babel.config.js
@@ -2,6 +2,10 @@ module.exports = function(api) {
 	api.cache(true);
 	return {
 		presets: ['babel-preset-expo'],
-		plugins: ['istanbul']
+		plugins: [
+			'istanbul', 
+			'@babel/plugin-proposal-export-namespace-from',
+			'react-native-reanimated/plugin'
+		]
 	};
 };
