+++ package.json
@@ -15,6 +15,9 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^0.17.2",
+    "@react-navigation/drawer": "^6.6.0",
+    "@react-navigation/native": "^6.1.4",
+    "@react-navigation/stack": "^6.3.14",
     "@types/react": "~18.0.24",
     "@types/react-native": "~0.70.6",
     "expo": "~47.0.12",
@@ -23,6 +26,9 @@
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
+    "react-native-gesture-handler": "~2.8.0",
+    "react-native-reanimated": "~2.12.0",
+    "react-native-safe-area-context": "4.4.1",
     "react-native-web": "~0.18.9",
     "typescript": "^4.6.3"
   },
