+++ src/screens/Home.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native';
+
+export function Home() {
+	return (
+		<View style={styles.container}>
+			<Text>Open up App.js to start working on your app!</Text>
+			<StatusBar style="auto" />
+		</View>
+	);
+}
+
+const styles = StyleSheet.create({
+	container: {
+		flex: 1,
+		backgroundColor: '#fff',
+		alignItems: 'center',
+		justifyContent: 'center',
+	},
+});
